@@ -3,12 +3,14 @@ package br.com.poc.factory.pattern.dto;
 import com.google.gson.Gson;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class ComputerDTO implements Serializable {
     private String name;
     private int ramSize;
     private int hardDiskSize;
     private String processorName;
+    private BigDecimal discount;
 
     public String getName() {
         return name;
@@ -19,15 +21,15 @@ public class ComputerDTO implements Serializable {
     }
 
     public int getRamSize() {
-        return ramSize;
+        return this.ramSize;
     }
 
     public void setRamSize(int ramSize) {
         this.ramSize = ramSize;
     }
 
-    public int getHardDisSize() {
-        return hardDiskSize;
+    public int getHardDiskSize() {
+        return this.hardDiskSize;
     }
 
     public void setHardDisSize(int hardDiskSize) {
@@ -35,11 +37,19 @@ public class ComputerDTO implements Serializable {
     }
 
     public String getProcessorName() {
-        return processorName;
+        return this.processorName;
     }
 
     public void setProcessorName(String processorName) {
         this.processorName = processorName;
+    }
+
+    public BigDecimal getDiscount() {
+        return this.discount;
+    }
+
+    public void setDiscount(BigDecimal discount) {
+        this.discount = discount;
     }
 
     @Override
