@@ -8,6 +8,6 @@ public abstract  class Product {
     public void sellProduct() {
         IProduct product = this.createProduct();
         System.out.println("Selling product: " + product.getProductDetails().toString());
-        product.sellProduct();
+        product.sell();
     }
 }
