@@ -4,6 +4,7 @@ import br.com.poc.factory.pattern.dto.ComputerDTO;
 import br.com.poc.factory.pattern.interfaces.IProduct;
 
 import javax.swing.*;
+import java.math.BigDecimal;
 
 public class Computer implements IProduct {
     private ComputerDTO computerDTO;
@@ -13,7 +14,7 @@ public class Computer implements IProduct {
     }
 
     @Override
-    public void sellProduct() {
+    public void sell() {
         JOptionPane.showMessageDialog(null, "Product Sold: " + this.getProductDetails(), "Factory Pattern", JOptionPane.INFORMATION_MESSAGE);
     }
 
@@ -28,6 +29,7 @@ public class Computer implements IProduct {
         this.computerDTO.setRamSize(16000);
         this.computerDTO.setProcessorName("INTEL");
         this.computerDTO.setHardDisSize(2048);
+        this.computerDTO.setDiscount(new BigDecimal(39));
     }
 }
 
